requirements.txt
python-vlc==3.0.11115
pytest==6.2.2